advanced-api-project-1/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.1.0